+++ templates/base.html
@@ -17,7 +17,7 @@
   
         <nav class="d-inline-flex mt-2 mt-md-0 ms-md-auto">
           <a class="me-3 py-2 text-dark text-decoration-none" href="/report">Отчёт</a>
-          <a class="me-3 py-2 text-dark text-decoration-none" href="/about">Добавить задачу</a>
+          <a class="me-3 py-2 text-dark text-decoration-none" href="/about/">Добавить задачу</a>
           <a class="me-3 py-2 text-dark text-decoration-none" href="/test_page">Тест кнопка</a>
           
         </nav>
